test/web-platform-tests/to-upstream/html/webappapis/timers/settimeout-setinterval-handles.html

<!doctype html>
<title>Handles returned by setTimeout and setInterval</title>
<link rel="author" title="Alvaro Dias" href="mailto:alvarorahul@gmail.com">
<link rel="help" href="https://html.spec.whatwg.org/multipage/#dom-windowtimers-settimeout">
<link rel="help" href="https://heycam.github.io/webidl/#es-operations">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
"use strict";

async_test(t => {
  const a = setTimeout(() => {
    // shouldn't execute since we clear immediately
    assert_true(false);
  }, 50);
  clearTimeout(a);

  let c;

  const b = setTimeout(() => {
    // this callback should execute since we haven't cleared the timeout
    const d = setInterval(() => {
      clearInterval(d);
      t.done();
    }, 50);

    assert_equals(d, c + 1); // since we created the interval after the handle c
  }, 50);
  assert_equals(b, a + 1);

  c = setInterval(() => {
    // shouldn't execute since we clear immediately
    assert_true(false);
  }, 50);
  clearInterval(c);
  assert_equals(c, b + 1);

}, "setTimeout and setInterval should return numeric handles which increment");
</script>
